+++ src/gateways/YouTube.php
@@ -192,7 +192,7 @@ class YouTube extends Gateway
      *
      * @since 2.0.0
      */
-    public function getEmbedFormat(): string
+    public function getEmbedUrlFormat(): string
     {
         return 'https://www.youtube.com/embed/%s?wmode=transparent';
     }