+++ CHANGELOG.md
@@ -84,6 +84,7 @@ Changelog
 - Moved `dukt\videos\base\Gateway::getAccount()` to `dukt\videos\base\Gateway::getOauthAccount()`
 - Changed `dukt\videos\base\Gateway::getVideoByUrl()` is now final
 - Changed `dukt\videos\base\Gateway::getVideoById()` is now final ; use `dukt\videos\base\Gateway::fetchVideoById()` to get video from the gateway API
+- Moved `dukt\videos\base\Gateway::getEmbedFormat()` to `dukt\videos\base\Gateway::getEmbedUrlFormat()`
 - Moved `dukt\videos\base\Gateway::getEmbed()` to `dukt\videos\helpers\EmbedHelper::getEmbedHtml()` and need `dukt\videos\model\Video` as params (instead of video id)
 - Moved `dukt\videos\base\Gateway::getEmbedUrl()`  to `dukt\videos\helpers\EmbedHelper::getEmbedUrl()` and need `dukt\videos\model\Video` as params (instead of video id)
 - Changed `dukt\videos\base\Gateway::getVideos()` is now final